Move the game counters ECS type defs to the interfaces crate. Prep to hoist the db to another crate
