+++ crates/thetawave_interface/src/lib.rs
@@ -1,2 +1,3 @@
+pub mod game;
 pub mod spawnable;
 pub mod states;
